% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowUnmix.R
\name{flowUnmix}
\alias{flowUnmix}
\title{Unmix your flow cytometry files}
\usage{
flowUnmix(
  fs,
  cs,
  unstained = NULL,
  unmixMethod = "lsfit",
  multiplier = 10000,
  spectrum = FALSE
)
}
\arguments{
\item{fs}{A flowSet of files to unmix}

\item{cs}{A flowset of single colour controls}

\item{unstained}{A flowframe of the negative control (optional)}

\item{unmixMethod}{Choose an inbuilt unmixing method: lsfit (the default), ginv, qr.solve ,lm.fit, crosspod, nnls, baselm}

\item{multiplier}{Most of these methods produce 0-1 values.  Multiply these to make them more palatable to your eye (default=10000)}

\item{spectrum}{TRUE uses flowSpectrum to output spectral plots}
}
\value{
fcs file and optionally spectrum images
}
\description{
flowUnmix requires you to input a flowset of single colour controls, optionally a negative flowframe, and a flowset of samples.
It then unmixes this data and outputs a new fcs file.  Various unmixing methods are available.
}
\details{
This function is the wrapper function for unmix_ff() and controlData()
}
