% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popchooser.R
\name{popChooser}
\alias{popChooser}
\title{Selecting the positive events in the unmixing controls}
\usage{
popChooser(singlePositives, negative, popCheck)
}
\arguments{
\item{singlePositives}{A flowset of single colour controls}

\item{negative}{A flowframe of the negative control}

\item{popCheck}{Prints to R a normalised spectrum of the chosen events}
}
\value{
median fluorescence intensity of each parameter in each single colour control
}
\description{
flowUnmix requires you to input a flowset of single colour controls, optionally a negative flowframe, and a flowset of samples.
It then unmixes this data and outputs a new fcs file.  Various unmixing methods are available.
}
\details{
popChooser() tries to guess the positive population and optionally prints the normalised spectrum of the chosen population.
}
\seealso{
\code{\link[flowUnmix]{controlData}},
\code{\link[flowUnmix]{unmix_ff}},
\code{\link[flowUnmix]{flowUnmix}}
}
\author{
Christopher Hall, Babraham Institute
}
