% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmix_ff.R
\name{unmix_ff}
\alias{unmix_ff}
\title{core unmixing functionality}
\usage{
unmix_ff(fs, control, unstainedsamp, unmixMethod, multiplier)
}
\arguments{
\item{fs}{A flowSet of files to unmix}

\item{control}{Matrix of the control data}

\item{unstainedsamp}{A flowframe of the negative control of the sample if different from unstainedctrl (optional)}

\item{unmixMethod}{Choose an inbuilt unmixing method: lsfit (the default), ginv, qr.solve ,lm.fit, crosspod, nnls, baselm}

\item{multiplier}{Most of these methods produce 0-1 values.  Multiply these to make them more palatable to your eye (default=10000)}

\item{negative}{A flowframe of the negative control (optional)}
}
\value{
unmixed fcs file
}
\description{
flowUnmix requires you to input a flowset of single colour controls, optionally a negative flowframe, and a flowset of samples.
It then unmixes this data and outputs a new fcs file.  Various unmixing methods are available.
}
\details{
unmix_ff() is the main looping function that unmixes the data
}
\seealso{
\code{\link[flowUnmix]{controlData}},
\code{\link[flowUnmix]{flowUnmix}},
\code{\link[flowUnmix]{popchooser}}
}
\author{
Christopher Hall, Babraham Institute
}
