% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlData.R
\name{controlData}
\alias{controlData}
\title{Formatting the control data}
\usage{
controlData(cs, unstained)
}
\arguments{
\item{cs}{A flowset of single colour controls}

\item{unstained}{A flowframe of the negative control (optional)}
}
\value{
fcs file
}
\description{
flowUnmix requires you to input a flowset of single colour controls, optionally a negative flowframe, and a flowset of samples.
It then unmixes this data and outputs a new fcs file.  Various unmixing methods are available.
}
\details{
controlData() creates data matrix of the control data for importing into unmix_ff()
}
